conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=py37h06a4308_0
addict=2.4.0=pypi_0
aiohttp=3.7.4=py37h27cfd23_1
apex=0.1=pypi_0
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
brotlipy=0.7.0=py37h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py37h06a4308_2
cffi=1.14.5=py37h261ae71_0
chardet=3.0.4=py37h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py37_0
cython=0.29.23=py37h2531618_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.7.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.6.0=pyhd3eb1b0_0
glib=2.68.2=h36276a3_0
google-auth=1.30.1=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imagedegrade=0.2.3=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py37h06a4308_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py37h27cfd23_1
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.1=py37ha9443f7_2
mmcv-full=1.3.8=pypi_0
mmdet=2.13.0=dev_0
mmpycocotools=12.0.3=pypi_0
multidict=5.1.0=py37h27cfd23_2
ncurses=6.2=he6710b0_1
networkx=2.6.1=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
numpy=1.20.2=py37h2d18471_0
numpy-base=1.20.2=py37hfae3a4d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv-python=4.5.2.54=pypi_0
openssl=1.1.1m=h7f8727e_0
pandas=1.2.5=py37h295c915_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pillow=8.2.0=py37he98fc37_0
pip=21.1.1=py37h06a4308_0
protobuf=3.14.0=py37h2531618_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycococreatortools=0.2.0=pypi_0
pycocotools=2.0=pypi_0
pycparser=2.20=py_2
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.7.0=py3.7_cuda11.0.221_cudnn8.0.3_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.1=py37ha9443f7_0
scipy=1.6.2=py37had2a1c9_1
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.35.4=hdfb4753_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
terminaltables=3.1.0=pypi_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.0=py37
torchvision=0.8.0=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.31.0=pypi_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
